requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-core
langchain-openai
python-dotenv
requests